PA-3360: Extracted TreeNodesFactory from ObjectExplorer
