+++ Src/SwqlStudio/ObjectExplorer/ObjectExplorer.cs
@@ -304,11 +304,7 @@ namespace SwqlStudio
 
                 foreach (var arg in FindProvider(e.Node).GetVerbArguments(verb))
                 {
-                    string text = $"{arg.Name} ({arg.Type})";
-                    var argNode = new TreeNode(text) { SelectedImageKey = "Argument" };
-                    argNode.ImageKey = argNode.SelectedImageKey;
-                    argNode.Tag = arg;
-                    argNode.ToolTipText = ToolTipBuilder.ToToolTip(arg);
+                    var argNode = TreeNodesFactory.CreateVerbArgumentNode(arg);
                     e.Node.Nodes.Add(argNode);
                 }
             }
@@ -558,7 +554,7 @@ namespace SwqlStudio
             _serverContextMenuItems.Add(connection.Title, serverContextMenu);
 
 
-            TreeNode node = CreateDatabaseNode(provider, connection);
+            TreeNode node = TreeNodesFactory.CreateDatabaseNode(provider, connection);
 
             TreeNode[] existingNodes = _treeData.Nodes.Find(node.Name, false);
             if (existingNodes.Length == 0)
@@ -616,17 +612,6 @@ namespace SwqlStudio
             serverContextMenu.Dispose();
         }
 
-        private static TreeNode CreateDatabaseNode(IMetadataProvider provider, ConnectionInfo connection)
-        {
-            TreeNode node = new TreeNodeWithConnectionInfo(provider.Name, connection);
-            node.SelectedImageKey = "Database";
-            node.ImageKey = "Database";
-            node.Tag = provider;
-            node.Name = node.Text;
-
-            return node;
-        }
-
         private void AddTablesToNode(TreeNode parent, IMetadataProvider provider, ConnectionInfo connection)
         {
             switch (EntityGroupingMode)
@@ -638,16 +623,8 @@ namespace SwqlStudio
                     foreach (var group in provider.Tables.GroupBy(e => e.Namespace).OrderBy(g => g.Key))
                     {
                         TreeNode[] childNodes = MakeEntityTreeNodes(provider, connection, group.OrderBy(e => e.FullName));
-
-                        int countChilds = childNodes.Length;
-                        var namespaceNode = new TreeNode(string.Format("{0} ({1} item{2})", group.Key, countChilds, countChilds > 1 ? "s" : string.Empty))
-                        {
-                            Tag = group.Key,
-                            ImageKey = "Namespace"
-                        };
-                        namespaceNode.SelectedImageKey = namespaceNode.ImageKey;
-
-                        namespaceNode.Nodes.AddRange(childNodes);
+                        var namespaceName = group.Key;
+                        var namespaceNode = TreeNodesFactory.CreateNamespaceNode(childNodes, namespaceName);
                         parent.Nodes.Add(namespaceNode);
                     }
                     break;
@@ -657,19 +634,8 @@ namespace SwqlStudio
                         (key, group) => new { Key = key, Entities = group }).OrderBy(item => item.Key.FullName))
                     {
                         TreeNode[] childNodes = MakeEntityTreeNodes(provider, connection, group.Entities.OrderBy(e => e.FullName));
-
-                        int countChilds = childNodes.Length;
-                        var baseTypeNode = new TreeNodeWithConnectionInfo(
-                            string.Format("{0} ({1} item{2})", group.Key.FullName, countChilds, countChilds > 1 ? "s" : string.Empty),
-                            connection)
-                        {
-                            Tag = group.Key
-                        };
-                        baseTypeNode.ImageKey = !group.Key.IsAbstract ? "BaseType" : "BaseTypeAbstract";
-                        baseTypeNode.SelectedImageKey = baseTypeNode.ImageKey;
-
-                        baseTypeNode.Nodes.AddRange(childNodes);
-                        parent.Nodes.Add(baseTypeNode);
+                        var entityNode = TreeNodesFactory.CreateEntityNode(connection, @group.Key, childNodes);
+                        parent.Nodes.Add(entityNode);
                     }
                     break;
                 case SwqlStudio.EntityGroupingMode.ByHierarchy:
@@ -716,97 +682,7 @@ namespace SwqlStudio
 
         private static TreeNode[] MakeEntityTreeNodes(IMetadataProvider provider, ConnectionInfo connection, IEnumerable<Entity> entities)
         {
-            return entities.Select(e => MakeEntityTreeNode(provider, connection, e)).ToArray();
-        }
-
-        private static TreeNodeWithConnectionInfo MakeEntityTreeNode(IMetadataProvider provider, ConnectionInfo connection, Entity entity)
-        {
-            var node = new TreeNodeWithConnectionInfo(entity.FullName, connection);
-            node.ImageKey = GetImageKey(entity);
-            node.SelectedImageKey = node.ImageKey;
-            node.Tag = entity;
-            node.ToolTipText = ToolTipBuilder.ToToolTip(connection, entity);
-
-            // Add keys
-            AddPropertiesToNode(node, entity.Properties.Where(c => c.IsKey));
-
-            // Add the simple Properties
-            AddPropertiesToNode(node, entity.Properties.Where(c => !c.IsInherited && !c.IsNavigable && !c.IsKey));
-
-            // Add the inherited Properties
-            AddPropertiesToNode(node, entity.Properties.Where(c => c.IsInherited && !c.IsNavigable && !c.IsKey));
-
-            // Add the Navigation Properties
-            AddPropertiesToNode(node, entity.Properties.Where(c => c.IsNavigable));
-
-            AddVerbsToNode(node, entity, provider);
-            return node;
-        }
-
-        private static string GetImageKey(Entity table)
-        {
-            if (table.IsIndication)
-                return "Indication";
-            else if (table.IsAbstract)
-                return "TableAbstract";
-            else if (table.CanCreate || table.CanDelete || table.CanUpdate)
-                return "TableCrud";
-
-            return "Table";
-        }
-
-        private static void AddVerbsToNode(TreeNode parent, Entity table, IMetadataProvider provider)
-        {
-            foreach (var verb in table.Verbs.OrderBy(v => v.Name))
-            {
-                TreeNode verbNode = new TreeNode(verb.Name);
-                verbNode.SelectedImageKey = "Verb";
-                verbNode.ImageKey = verbNode.SelectedImageKey;
-                verbNode.Tag = verb;
-                verbNode.ToolTipText = ToolTipBuilder.ToToolTip(verb);
-
-                parent.Nodes.Add(verbNode);
-
-                verbNode.Nodes.Add(new ObjectExplorer.ArgumentsPlaceholderTreeNode(verb, provider));
-            }
-        }
-
-        private class ArgumentsPlaceholderTreeNode : TreeNode
-        {
-            public Verb Verb { get; set; }
-            public IMetadataProvider Provider { get; set; }
-
-            public ArgumentsPlaceholderTreeNode(Verb verb, IMetadataProvider provider)
-            {
-                Verb = verb;
-                Provider = provider;
-            }
-        }
-
-        private static void AddPropertiesToNode(TreeNode parent, IEnumerable<Property> properties)
-        {
-            foreach (Property property in properties.OrderBy(c => c.Name))
-            {
-                string name = $"{property.Name} ({property.Type})";
-                TreeNode node = new TreeNode(name);
-                node.SelectedImageKey = GetColumnIcon(property);
-                node.ImageKey = node.SelectedImageKey;
-                node.Tag = property;
-                node.ToolTipText = ToolTipBuilder.ToToolTip(property, name);
-                parent.Nodes.Add(node);
-            }
-        }
-
-        private static string GetColumnIcon(Property column)
-        {
-            if (column.IsNavigable)
-                return "Link";
-            if (column.IsKey)
-                return "KeyColumn";
-            if (column.IsInherited)
-                return "InheritedColumn";
-
-            return "Column";
+            return entities.Select(e => TreeNodesFactory.MakeEntityTreeNode(provider, connection, e)).ToArray();
         }
 
         private void Crud(Entity entity, CrudOperation operation)
